Removing builder structs from public API of casper-types
