+++ execution_engine_testing/tests/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod lmdb_fixture;
 pub mod wasm_utils;
+pub use casper_engine_test_support::genesis_config_builder::GenesisConfigBuilder;
 
 #[cfg(test)]
 mod test;
