+++ node/src/reactor/main_reactor/tests/binary_port.rs
@@ -30,8 +30,8 @@ use casper_types::{
     CLValue, CLValueDictionary, ChainspecRawBytes, Contract, ContractWasm, ContractWasmHash,
     DictionaryAddr, Digest, EntityAddr, EntityKind, EntityVersions, GlobalStateIdentifier, Key,
     KeyTag, NextUpgrade, Package, PackageAddr, PackageHash, Peers, ProtocolVersion, PublicKey,
-    Rewards, SecretKey, SignedBlock, StoredValue, Transaction, TransactionRuntime,
-    TransactionV1Builder, Transfer, URef, U512,
+    Rewards, SecretKey, SignedBlock, StoredValue, Transaction, TransactionRuntime, Transfer, URef,
+    U512,
 };
 use futures::{SinkExt, StreamExt};
 use rand::Rng;
@@ -43,7 +43,7 @@ use crate::{
     testing::{
         self, filter_reactor::FilterReactor, network::TestingNetwork, ConditionCheckReactor,
     },
-    types::NodeId,
+    types::{transaction::transaction_v1_builder::TransactionV1Builder, NodeId},
 };
 
 use super::{InitialStakes, TestFixture, ERA_ONE};
