+++ node/src/types/transaction.rs
@@ -1,6 +1,10 @@
+pub(crate) mod arg_handling;
 mod deploy;
 mod meta_transaction;
 mod transaction_footprint;
 pub(crate) use deploy::LegacyDeploy;
 pub(crate) use meta_transaction::{MetaTransaction, TransactionHeader};
 pub(crate) use transaction_footprint::TransactionFootprint;
+pub(crate) mod fields_container;
+pub(crate) mod initiator_addr_and_secret_key;
+pub(crate) mod transaction_v1_builder;
