+++ node/src/types/transaction/meta_transaction/meta_transaction_v1.rs
@@ -1,8 +1,9 @@
 use super::transaction_lane::{calculate_transaction_lane, TransactionLane};
+use crate::types::transaction::arg_handling;
 use casper_types::{
-    arg_handling, bytesrepr::ToBytes, crypto, Approval, Chainspec, Digest, DisplayIter, Gas,
-    HashAddr, InitiatorAddr, InvalidTransaction, InvalidTransactionV1, PricingHandling,
-    PricingMode, TimeDiff, Timestamp, TransactionArgs, TransactionConfig, TransactionEntryPoint,
+    bytesrepr::ToBytes, crypto, Approval, Chainspec, Digest, DisplayIter, Gas, HashAddr,
+    InitiatorAddr, InvalidTransaction, InvalidTransactionV1, PricingHandling, PricingMode,
+    TimeDiff, Timestamp, TransactionArgs, TransactionConfig, TransactionEntryPoint,
     TransactionRuntime, TransactionScheduling, TransactionTarget, TransactionV1,
     TransactionV1ExcessiveSizeError, TransactionV1Hash, U512,
 };
